# MIS.Api/Controllers/Contract/src/OpenApi/Controllers/DiseasesApi.cs
/*
 * Medical Information System (MIS) API
 *
 * API для медицинской информационной системы
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Mvc;
using MIS.Api.Controllers.Contract.OpenApi.Attributes;
using MIS.Api.Controllers.Contract.OpenApi.Models;
using Swashbuckle.AspNetCore.Annotations;

namespace MIS.Api.Controllers.Contract.OpenApi.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class DiseasesApiController : ControllerBase
    { 
        /// <summary>
        /// Получить общий список болезней
        /// </summary>
        /// <remarks>Возвращает список всех болезней в системе</remarks>
        /// <param name="page">Номер страницы (начиная с 1)</param>
        /// <param name="pageSize">Размер страницы</param>
        /// <param name="isChronic">Фильтр по хроническим болезням</param>
        /// <param name="isInfectious">Фильтр по инфекционным болезням</param>
        /// <response code="200">Успешный ответ со списком болезней</response>
        /// <response code="400">Некорректный запрос</response>
        /// <response code="500">Внутренняя ошибка сервера</response>
        [HttpGet]
        [Route("/diseases")]
        [ValidateModelState]
        [SwaggerOperation("GetAllDiseases")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetAllDiseases200Response), description: "Успешный ответ со списком болезней")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Некорректный запрос")]
        [SwaggerResponse(statusCode: 500, type: typeof(Error), description: "Внутренняя ошибка сервера")]
        public abstract Task<IActionResult> GetAllDiseases([FromQuery (Name = "page")]int? page, [FromQuery (Name = "pageSize")][Range(1, 100)]int? pageSize, [FromQuery (Name = "isChronic")]bool? isChronic, [FromQuery (Name = "isInfectious")]bool? isInfectious);
    }
}