# MIS.Api/Controllers/Contract/src/OpenApi/Controllers/DoctorsApi.cs
/*
 * Medical Information System (MIS) API
 *
 * API для медицинской информационной системы
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Mvc;
using MIS.Api.Controllers.Contract.OpenApi.Attributes;
using MIS.Api.Controllers.Contract.OpenApi.Models;
using Swashbuckle.AspNetCore.Annotations;

namespace MIS.Api.Controllers.Contract.OpenApi.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class DoctorsApiController : ControllerBase
    { 
        /// <summary>
        /// Получить список докторов по специальности
        /// </summary>
        /// <remarks>Возвращает список докторов, отфильтрованных по специальности</remarks>
        /// <param name="specialty">Специальность доктора</param>
        /// <param name="page">Номер страницы (начиная с 1)</param>
        /// <param name="pageSize">Размер страницы</param>
        /// <response code="200">Успешный ответ со списком докторов</response>
        /// <response code="400">Некорректный запрос</response>
        /// <response code="500">Внутренняя ошибка сервера</response>
        [HttpGet]
        [Route("/doctors/by-specialty")]
        [ValidateModelState]
        [SwaggerOperation("GetDoctorsBySpecialty")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetDoctorsBySpecialty200Response), description: "Успешный ответ со списком докторов")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "Некорректный запрос")]
        [SwaggerResponse(statusCode: 500, type: typeof(Error), description: "Внутренняя ошибка сервера")]
        public abstract Task<IActionResult> GetDoctorsBySpecialty([FromQuery (Name = "specialty")][Required()]string specialty, [FromQuery (Name = "page")]int? page, [FromQuery (Name = "pageSize")][Range(1, 100)]int? pageSize);
    }
}